Change format to use one row per software, with multiple bits each.
Calico reserves 16 bits for itself, which would be ridiculously unwieldy
in the previous format. The original goal of listing individual bits was
so that we could show conflicts easily, by listing several pieces of
software for a bit.

Instead, this change adds a new "conflicts" section that highlights which
of the bit ranges from the prior section represent conflicts.

Signed-off-by: David Anderson <dave@natulte.net>